requirements.txt
numpy==2.1.3
scipy==1.15.1
scikit-learn==1.6.1
librosa==0.11.0
soundfile==0.13.1
torch==2.6.0
torchaudio==2.6.0
tensorflow==2.19.0
tqdm
coqui-tts==0.26.1
transformers==4.51.3
matplotlib==3.10.1